feat: Add uncategorized contact list
and some ContactsList refactor
