+++ src/components/ContactsList/ContactsList.jsx
@@ -5,11 +5,9 @@ import flag from 'cozy-flags'
 import Button from 'cozy-ui/transpiled/react/Button'
 import { translate } from 'cozy-ui/transpiled/react/I18n'
 
-import { categorizeContacts } from '../../helpers/contactList'
 import ContactsEmptyList from './ContactsEmptyList'
-import ContactRow from './ContactRow'
-import ContactHeaderRow from './ContactHeaderRow'
-
+import CategorizedList from './CategorizedList'
+import UncategorizedList from './UncategorizedList'
 import withSelection from '../Selection/selectionContainer'
 
 class ContactsList extends Component {
@@ -18,49 +16,35 @@ class ContactsList extends Component {
 
     if (contacts.length === 0) {
       return <ContactsEmptyList />
-    } else {
-      const isAllContactsSelected = contacts.length === selection.length
-      const categorizedContacts = categorizeContacts(contacts, t('empty-list'))
-
-      return (
-        <div className="list-wrapper">
-          {flag('select-all-contacts') && (
-            <div>
-              <Button
-                label={
-                  isAllContactsSelected ? t('unselect-all') : t('select-all')
-                }
-                theme="secondary"
-                onClick={() =>
-                  isAllContactsSelected ? clearSelection() : selectAll(contacts)
-                }
-              />
-            </div>
-          )}
-          <ol className="list-contact">
-            {Object.entries(categorizedContacts).map(([header, contacts]) => (
-              <li key={`cat-${header}`}>
-                <ContactHeaderRow key={header} header={header} />
-                <ol className="sublist-contact">
-                  {contacts.map(contact => (
-                    <li key={`contact-${contact._id}`}>
-                      <ContactRow
-                        id={contact._id}
-                        key={contact._id}
-                        contact={contact}
-                      />
-                    </li>
-                  ))}
-                </ol>
-              </li>
-            ))}
-          </ol>
-          <div />
-        </div>
-      )
     }
+
+  const isSearched = false // TODO use context to determine bool result according to search input
+  const List = isSearched ? UncategorizedList : CategorizedList
+  const isAllContactsSelected = contacts.length === selection.length
+
+    const handleAllContactSelection = () => {
+      isAllContactsSelected ? clearSelection() : selectAll(contacts)
+    }
+
+    return (
+      <div className="list-wrapper">
+        {flag('select-all-contacts') && (
+          <div>
+            <Button
+              label={
+                isAllContactsSelected ? t('unselect-all') : t('select-all')
+              }
+              theme="secondary"
+              onClick={handleAllContactSelection}
+            />
+          </div>
+        )}
+        <List contacts={contacts} />
+      </div>
+    )
   }
 }
+
 ContactsList.propTypes = {
   contacts: PropTypes.array.isRequired
 }
